# Conflicts:
#	Vista/src/main/java/helper/RentaHelper.java
#	Vista/src/main/java/ui/RentaBeanUI.java
#	Vista/src/main/webapp/Rentas.xhtml
#	negocio/src/main/java/mx/desarollo/delegate/DelegateRenta.java
#	negocio/src/main/java/mx/desarollo/facade/FacadeRenta.java
#	persistencia/src/main/java/mx/avanti/desarollo/dao/RentaDAO.java
#	persistencia/src/main/resources/META-INF/persistence.xml

+++ entidad/src/main/java/mx/desarollo/entity/StockReservadoDiario.java
@@ -26,7 +26,7 @@ public class StockReservadoDiario {
     private LocalDate fecha;
 
     @Column(name = "cantidad_reservada", columnDefinition = "int UNSIGNED not null")
-    private Long cantidadReservada;
+    private Integer cantidadReservada;
 
     public Long getId() {
         return id;
@@ -52,11 +52,11 @@ public class StockReservadoDiario {
         this.fecha = fecha;
     }
 
-    public Long getCantidadReservada() {
+    public Integer getCantidadReservada() {
         return cantidadReservada;
     }
 
-    public void setCantidadReservada(Long cantidadReservada) {
+    public void setCantidadReservada(Integer cantidadReservada) {
         this.cantidadReservada = cantidadReservada;
     }
